Fix bug with uWSGI
Global variables had been moved into the `__main__` execution. This
worked correctly when using Flask but break with uWSGI (which instead
of `__main__` "sees" `app`).

Reorganized code and called `dbReader` using a `with` statement. This
needs to be reviewed in the future, to see if there is a way to reuse
the object, with the assumption that creating an object has a non
trivial overhead (might not be true).